+++ example/spec/parallel/kinesisTests/LambdaEventSourceSpec.js
@@ -13,7 +13,7 @@ const {
   cleanupProviders,
   addCollections,
   cleanupCollections,
-  rulesList,
+  readJsonFilesFromDir,
   deleteRules
 } = require('@cumulus/integration-tests');
 const { randomString } = require('@cumulus/common/test-utils');
@@ -61,7 +61,7 @@ describe('When adding multiple rules that share a kinesis event stream', () => {
 
   async function cleanUp() {
     // delete rules
-    const rulesToDelete = await rulesList(testConfig.stackName, testConfig.bucket, ruleDirectory);
+    const rulesToDelete = await readJsonFilesFromDir(ruleDirectory);
     // clean up stack state added by test
     console.log(`\nCleaning up stack & deleting test stream '${streamName}'`);
     await deleteRules(testConfig.stackName, testConfig.bucket, rulesToDelete, ruleSuffix);